tests init
