+++ moonpay-api/package.json
@@ -6,7 +6,8 @@
     "scripts": {
         "start": "node index.js",
         "dev": "env NODE_ENV=development nodemon --exec npm start",
-        "test": "npm run lint && env NODE_ENV=development mocha --exit --require @babel/register test/*",
+        "test": "cross-env NODE_ENV=development LOG_DEBUG=no mocha --exit",
+        
         "lint": "standard --fix 'src/**/*.js'"
     },
     "repository": {
@@ -49,7 +50,10 @@
         "@babel/plugin-proposal-object-rest-spread": "^7.5.5",
         "@babel/preset-env": "^7.5.5",
         "@babel/register": "^7.5.5",
+        "axios": "^0.19.0",
         "chai": "^4.2.0",
+        "chai-http": "^4.3.0",
+        "cross-env": "^5.2.0",
         "mocha": "^6.2.0",
         "morgan": "^1.9.1",
         "node-cron": "^2.0.3",
